Use borrow trait to add items into sketch
Using the borrow trait when inserting a value into the sketch, i.e.
`value: &Q` with `H: Borrow<Q>` rather than `value: &H`, makes the
sketch types much more flexible. For example, you don't run into the
issue of having a type that contains a lifetime in the sketch:
`struct HyperLogLog<H>` with `H = MyType<'a>` causes issues when
`MyType<'a>` is a temporary value that references a type that's being
destroyed at the end of some scope.